requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.1
django-jazzmin==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
openapi-codec==1.3.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
